+++ src/main/java/com/assetflow/loan/dto/LoanCreateRequest.java
@@ -1,11 +1,13 @@
 package com.assetflow.loan.dto;
 
 import jakarta.validation.constraints.NotNull;
+import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 
 @Getter
 @RequiredArgsConstructor
+@AllArgsConstructor
 public class LoanCreateRequest {
     @NotNull
     private Long memberId;