+++ src/main/java/com/assetflow/reservation/dto/ReservationCreateRequest.java
@@ -7,6 +7,7 @@ import lombok.NoArgsConstructor;
 
 @Getter
 @NoArgsConstructor
+@AllArgsConstructor
 public class ReservationCreateRequest {
         @NotNull
         private Long memberId;